+++ app/main/forms.py
@@ -56,13 +56,9 @@ from app.main.validators import (
 )
 from app.models.feedback import PROBLEM_TICKET_TYPE, QUESTION_TICKET_TYPE
 from app.models.organisation import Organisation
-from app.models.roles_and_permissions import (
-    broadcast_permissions,
-    permissions,
-    roles,
-)
 from app.utils import merge_jsonlike
 from app.utils.user import distinct_email_addresses
+from app.utils.user_permissions import broadcast_permissions, permissions, roles
 
 
 def get_time_value_and_label(future_time):
