+++ app/main/views/manage_users.py
@@ -28,9 +28,9 @@ from app.main.forms import (
     PermissionsForm,
     SearchUsersForm,
 )
-from app.models.roles_and_permissions import broadcast_permissions, permissions
 from app.models.user import InvitedUser, User
 from app.utils.user import is_gov_user, user_has_permissions
+from app.utils.user_permissions import broadcast_permissions, permissions
 
 
 @main.route("/services/<uuid:service_id>/users")
